requirements.txt
pandas
plotly_express
streamlit
matplotlib
numpy
seaborn
plotly.graph_objects